myproject/requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
Django==5.1.1
djangorestframework==3.15.2
msgpack==1.1.0
redis==5.0.8
sqlparse==0.5.1
tzdata==2024.1
